WIP: Fix several bugs, refactoring
