+++ example/coinjump_play.py
@@ -8,7 +8,8 @@ sys.path.insert(0, '../')
 from nudge.utils import make_deterministic
 from nudge.environments.getout.getout.imageviewer import ImageViewer
 from nudge.environments.getout.getout.getout.helpers import create_getout_instance
-from nsfr.utils import get_nsfr_model, get_predictions
+from nsfr.utils import get_predictions
+from nsfr.common import get_nsfr_model
 from nudge.agents.utils_getout import extract_logic_state_getout
 
 KEY_SPACE = 32
