+++ example/loot_play.py
@@ -7,7 +7,8 @@ sys.path.insert(0, '../')
 
 from nudge.utils import make_deterministic
 from nudge.environments.procgen.procgen import ProcgenGym3Env
-from nsfr.utils import get_nsfr_model, get_predictions
+from nsfr.utils import get_predictions
+from nsfr.common import get_nsfr_model
 from nudge.agents.utils_loot import extract_logic_state_loot
