+++ nsfr/nsfr/data/lang/atari/freeway_bs_rf1/consts.txt
@@ -1,7 +1,4 @@
-ocar:obj2,obj3,obj4,obj5,obj6,obj7,obj8,obj9,obj10,obj11
-type:car
-image:img
-
 oagent:obj1
-type:chicken
+ocar:obj2,obj3,obj4,obj5,obj6,obj7,obj8,obj9,obj10,obj11
+type:car,chicken
 image:img
