+++ nudge/.gitignore
@@ -1,1 +1,2 @@
 *.build
+/nudge.egg-info
