+++ nudge/train.py
@@ -6,7 +6,7 @@ import pickle
 import sys
 import time
 
-import gym
+import gymnasium as gym
 import numpy as np
 
 # import wandb
